+++ lingui.config.ts
@@ -45,6 +45,14 @@ const config: LinguiConfig = {
       include: ['packages/toggle/src/**/*.{ts,tsx}'],
       path: 'packages/toggle/src/locales/{locale}/messages',
     },
+    {
+      include: ['packages/steps/src/**/*.{ts,tsx}'],
+      path: 'packages/steps/src/locales/{locale}/messages',
+    },
+    {
+      include: ['packages/attention/src/**/*.{ts,tsx}'],
+      path: 'packages/attention/src/locales/{locale}/messages',
+    },
   ],
   compileNamespace: 'es',
 };