+++ packages/attention/src/props.ts
@@ -2,12 +2,7 @@ import { MutableRefObject } from 'react';
 
 export type Directions = 'top' | 'right' | 'bottom' | 'left';
 
-export type ArrowProps = {
-  /**
-   * Opposite direction of which the arrow should point
-   */
-  direction: Directions;
-
+export type AttentionVariants = {
   /**
    * Render tooltip
    */
@@ -24,10 +19,22 @@ export type ArrowProps = {
   popover?: boolean;
 
   /**
+   * Render highlight
+   */
+  highlight?: boolean;
+}
+
+export type ArrowProps = {
+  /**
+   * Opposite direction of which the arrow should point
+   */
+  direction: Directions;
+
+  /**
    * Forward arrow ref so Attention element can use it
    */
   ref?: React.Ref<HTMLDivElement>;
-};
+} & AttentionVariants;
 
 export type AttentionProps = {
   /**
@@ -46,21 +53,15 @@ export type AttentionProps = {
    */
   children?: JSX.Element[] | JSX.Element;
   
-
-  /**
+ /**
    * Sets the aria-label attribute for the Attention component
    */
-  ariaLabel?: string;
-
+ 'aria-label'?: string;
+ 
  /**
    * Sets the role attribute for the Attention component
    */
- ariaRole?: string;
-
- /**
-   * Sets the aria-describedby attribute for the Attention component to associate it with the target element
-   */
- ariaDescribedby?: string;
+ role?: string;
  
   /**
    * Placement according to the target element
@@ -77,4 +78,19 @@ export type AttentionProps = {
    * Extend the Attention component container styling
    */
   className?: string;
-} & Omit<ArrowProps, 'direction'>;
+
+  /**
+   * id used for relationship ARIA attributes
+   */
+  id?: string;
+
+  /**
+   * Whether a close button is shown
+   */
+  canClose?: boolean;
+
+  /**
+   * Handler that is called when the close button is clicked
+   */
+  onDismiss?: () => void;
+} & AttentionVariants;